renamed name methods and class, delete handling errors
