+++ 3741/1/top_gems.rb
@@ -1,19 +1,18 @@
 require 'yaml'
 require 'optparse'
-require_relative 'lib/read_yaml'
+require_relative 'lib/yaml_reader'
 require_relative 'lib/gem_data_reader'
-require_relative 'lib/show_table'
+require_relative 'lib/gems_view'
 
 # main run class
 class TopGems
-  # :reek:NestedIterators
   def initialize
     @file_name = 'gems.yml'
-    OptionParser.new do |opts|
-      opts.on('--top=', '--top') { |value| @option_top = value.to_i }
-      opts.on('--name=', '--name') { |value| @option_name = value }
-      opts.on('--file=', '--file') { |value| @file_name = value }
-    end.parse!
+    option_parser = OptionParser.new
+    option_parser.on('--top=', '--top') { |value| @option_top = value.to_i }
+    option_parser.on('--name=', '--name') { |value| @option_name = value }
+    option_parser.on('--file=', '--file') { |value| @file_name = value }
+    option_parser.parse!
   end
 
   def render_table
@@ -24,7 +23,7 @@ class TopGems
   private
 
   def read_file
-    ReadYaml.call(@file_name, @option_name)
+    YamlReader.call(@file_name, @option_name)
   end
 end
